FINAL bug fixes
Если принципиально оценить именно FINAL commit, то оценивать нужно
предыдущий.

- Исправлена ошибка с кол-вом элементов ДНК
- Добавлен метод print для Cell
- Для рассчета кол-ва изменяемых элементов в mutate теперь используется
float вместо int

+++ DNA/DNA/main.m
@@ -20,7 +20,7 @@
     NSLog(@"Mutating for %i percents", percent);
     // В нашем конкретном случае можно было бы и не вычислять проценты, у нас
     // как раз 100 элементов.
-    int count = round([self.DNA count] / 100 * percent);
+    int count = round((float)([self.DNA count] / 100.0) * percent);
     
     // Массив с индексами измененных элементов ДНК
     NSMutableArray *changed = [[NSMutableArray alloc] init];
